pyproject.toml
[tool.poetry]
name = "nutrition-tracker"
version = "0.1.0"
description = ""
authors = ["SMassalski <stanislaw.massalski@gmail.com>"]
readme = "README.md"
packages = [{include = "nutrition_tracker", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.2"
django-environ = "^0.9.0"
django-rest-framework = "^0.1.0"

[tool.poetry.group.test.dependencies]
pytest-django = "^4.5.2"
pytest = "^7.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
black = "^22.10.0"
coverage = "^7.2.0"
pytest-cov = "^4.0.0"
docstr-coverage = "^2.2.0"
django-debug-toolbar = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"


[tool.pytest.ini_options]
django_find_project = false
pythonpath = [".", "src/nutrition_tracker"]
DJANGO_SETTINGS_MODULE = "nutrition_tracker.settings.test"
addopts = "--no-migrations --cov --cov-report html"


[tool.coverage.run]
branch = true
data_file = "coverage/.coverage"
source = ["src"]
omit = ["*/manage.py", "*/migrations/*", "src/nutrition_tracker/nutrition_tracker/*"]

[tool.coverage.report]
skip_empty = true
fail_under = 80
exclude_also = ["def __str__"]

[tool.coverage.html]
directory = "coverage/html/"
title = "Nutrition Tracker Test Coverage Report"
